Added test cases for backend

+++ PackYourBags_Backend/src/main/java/com/example/packyourbags_backend/controllers/UserController.java
@@ -30,12 +30,12 @@ public class UserController {
     }
 
     @GetMapping("/{id}")
-    public User getUser(@PathVariable Long id) {
+    public User getUser(@PathVariable int id) {
         return service.getUser(id);
     }
 
     @DeleteMapping("/{id}")
-    public void deleteUser(@PathVariable Long id) {
+    public void deleteUser(@PathVariable int id) {
         service.deleteUser(id);
     }
